+++ android/app/src/main/java/com/xttech/app/MainActivity.java
@@ -8,6 +8,7 @@ import com.getcapacitor.BridgeActivity;
 public class MainActivity extends BridgeActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(NativeTrackingPlugin.class);
         super.onCreate(savedInstanceState);
         CookieManager cookieManager = CookieManager.getInstance();
         cookieManager.setAcceptCookie(true);